+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+# dev
+
+* Fix: `pe.data_directories` is now correctly an array and not a structure.
+
 # v2.4.0 (2019-02-27)
 
 * Enhancement: Python interface of `String.pure_text` now returns `bytes` instead of `str` to prevent unicode decoding errors with strings containg invalid UTF-8 sequences.